+++ CHANGELOG.md
@@ -11,14 +11,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- New class `Redmine\Serializer\PathSerializer` to build an URL path with query parameters.
+- New class `Redmine\Serializer\JsonSerializer` to encode or normalize JSON data.
+- New class `Redmine\Serializer\XmlSerializer` to encode or normalize XML data.
 - Allow `Psr\Http\Message\RequestFactoryInterface` as Argument #2 ($requestFactory) in `Redmine\Client\Psr18Client::__construct()`
 - Added support for PHP 8.2
 
 ### Deprecated
 
 - Providing Argument #2 ($requestFactory) in `Redmine\Client\Psr18Client::__construct()` as type `Psr\Http\Message\ServerRequestFactoryInterface` is deprecated, provide as type `Psr\Http\Message\RequestFactoryInterface` instead
-- `Redmine\Api\AbstractApi::attachCustomFieldXML()` is deprecated
-- `Redmine\Api\Project::prepareParamsXml()` is deprecated
+- `Redmine\Api\AbstractApi::attachCustomFieldXML()` is deprecated, use `Redmine\Serializer\XmlSerializer::createFromArray()` instead
+- `Redmine\Api\Project::prepareParamsXml()` is deprecated, use `Redmine\Serializer\XmlSerializer::createFromArray()` instead
 
 ## [v2.2.0](https://github.com/kbsali/php-redmine-api/compare/v2.1.1...v2.2.0) - 2022-03-01